Rename teaser component storybook configuration to skip automatic snapshot creation
