+++ .storybook/register-components.js
@@ -1,5 +1,5 @@
 const components = [
-	require('../components/x-teaser/stories'),
+	require('../components/x-teaser/storybook'),
 	require('../components/x-increment/stories'),
 	require('../components/x-styling-demo/stories'),
 	require('../components/x-gift-article/stories'),
